Apply dhewm3 / 30850f8d9c5829ca553917b9d22e11de59c2bd16: Fix FS hashing for x86_64
git-svn-id: http://svn.thedarkmod.com/svn/darkmod_src/branches/x64@6208 49c82d7f-2e2a-0410-a16f-ae8f201b507f

+++ framework/FileSystem.cpp
@@ -305,7 +305,7 @@ private:
 
 private:
 	void					ReplaceSeparators( idStr &path, char sep = PATHSEPERATOR_CHAR );
-	long					HashFileName( const char *fname ) const;
+    int 					HashFileName(const char *fname) const;
 	int						ListOSFiles( const char *directory, const char *extension, idStrList &list );
 	FILE *					OpenOSFile( const char *name, const char *mode, idStr *caseSensitiveName = NULL );
 	FILE *					OpenOSFileCorrectName( idStr &path, const char *mode );
@@ -382,9 +382,9 @@ idFileSystemLocal::HashFileName
 return a hash value for the filename
 ================
 */
-long idFileSystemLocal::HashFileName( const char *fname ) const {
+int idFileSystemLocal::HashFileName( const char *fname ) const {
 	int		i;
-	long	hash;
+	int	hash;
 	char	letter;
 
 	hash = 0;
@@ -397,7 +397,7 @@ long idFileSystemLocal::HashFileName( const char *fname ) const {
 		if ( letter == '\\' ) {
 			letter = '/';		// damn path names
 		}
-		hash += (long)(letter * (i+119));
+		hash += (int)(letter * (i+119));
 		i++;
 	}
 	hash &= (FILE_HASH_SIZE-1);
@@ -824,7 +824,7 @@ bool idFileSystemLocal::FileIsInPAK( const char *relativePath ) {
 	searchpath_t	*search;
 	pack_t			*pak;
 	fileInPack_t	*pakFile;
-	long			hash;
+	int			hash;
 
 	if ( !searchPaths ) {
 		common->FatalError( "Filesystem call made without initialization\n" );
@@ -1129,7 +1129,7 @@ pack_t *idFileSystemLocal::LoadZipFile( const char *zipfile ) {
 	unz_global_info gi;
 	char			filename_inzip[MAX_ZIPPED_FILE_NAME];
 	unz_file_info	file_info;
-	long			hash;
+	int			hash;
 	int				fs_numHeaderLongs;
 	int *			fs_headerLongs;
 	FILE			*f;
@@ -2591,7 +2591,7 @@ idFile *idFileSystemLocal::OpenFileReadFlags( const char *relativePath, int sear
 	pack_t *		pak;
 	fileInPack_t *	pakFile;
 	directory_t *	dir;
-	long			hash;
+	int			hash;
 	FILE *			fp;
 	
 	if ( !searchPaths ) {
